requirements.txt
google-generativeai
google-ai-generativelanguage
langchain
streamlit
python-dotenv
PyPDF2~=2.0
langchain-google-genai
cygrpc
grpcio

-e .